FDG-5819 Made the custome nivo layers reusable
