+++ src/layouts/explainer/sections/federal-spending/spending-trends/total-spending-chart/total-spending-chart-helper.js
@@ -241,9 +241,6 @@ export const getMarkers = (width, selectedChartView, gdpValue, spendingValue) =>
       fontSize: width < pxToNumber(breakpointLg) ? fontSize_10 : fontSize_14,
     },
   };
-
-
-
   return selectedChartView === "percentageGdp"
     ? []
     : [
@@ -258,4 +255,89 @@ export const getMarkers = (width, selectedChartView, gdpValue, spendingValue) =>
           value: spendingValue+1,
         },
       ];
+};
+
+export const lineChartCustomSlices = ( props, groupMouseLeave, mouseMove ) => {
+  return (
+    <g data-testid="customSlices"
+      onMouseLeave={groupMouseLeave}
+    >
+      {props.slices.map(slice => (
+        <rect
+          x={slice.x0}
+          y={slice.y0}
+          tabIndex={0}
+          width={slice.width}
+          height={slice.height}
+          strokeWidth={0}
+          strokeOpacity={0.25}
+          fillOpacity={0}
+          onMouseEnter={() => props.setCurrentSlice(slice)}
+          onMouseMove={() => mouseMove(slice)}
+          onMouseLeave={() => props.setCurrentSlice(null)}
+        />
+      ))}
+    </g>
+  );
+};
+
+export const lineChartCustomPoints = props => {
+  const { currentSlice, borderWidth, borderColor, points } = props;
+
+  
+    const lastGdpPoints = points.filter(g => g.serieId == 'GDP').pop();
+
+    const currentSpendingPoint = currentSlice?.points?.length
+      ? currentSlice.points[0]
+      : points[points.length - 1];
+
+    const currentGdpPoint = currentSlice?.points?.length
+      ? currentSlice.points[1]
+      : lastGdpPoints;
+
+    return (
+      <g data-testid="customPoints">
+        <circle
+          fill={'#D8D8D8'}
+          r={8}
+          strokeWidth={borderWidth}
+          stroke={borderColor}
+          fillOpacity={0.35}
+          cx={currentSpendingPoint.x}
+          cy={currentSpendingPoint.y}
+        />
+        <circle
+          r={2}
+          strokeWidth={'4'}
+          stroke={'#000000'}
+          fill={'#000000'}
+          fillOpacity={0.85}
+          cx={currentSpendingPoint.x}
+          cy={currentSpendingPoint.y}
+        />
+        {currentGdpPoint && (
+          <>
+            <circle
+              fill={'#D8D8D8'}
+              r={8}
+              strokeWidth={borderWidth}
+              stroke={borderColor}
+              fillOpacity={0.35}
+              cx={currentGdpPoint.x}
+              cy={currentGdpPoint.y}
+            />
+            <circle
+              r={2}
+              strokeWidth={'4'}
+              stroke={'#000000'}
+              fill={'#000000'}
+              fillOpacity={0.85}
+              cx={currentGdpPoint.x}
+              cy={currentGdpPoint.y}
+            />
+          </>
+        )}
+      </g>
+    );
+  
 };
